+++ src/File.php
@@ -56,7 +56,7 @@ class File
             throw new \Exception('Cant add row with different fieldset');
         }
 
-        $this->data = $row->__toString() . "\r\n";
+        $this->data = $row->render() . "\r\n";
     }
 
     /**