+++ CHANGELOG.md
@@ -30,6 +30,7 @@
 * Added `File::createRow` method.
 * Added `Row::set` and `Row::get` methods. For performance, these should be favoured over the magic `__get` and `__set` methods.
 * Added `Frameworks` directory with framework-specific utilities.
+* Added `Row::render` method
 
 ### Removed